Giant commit (20 days of work).
Rigidbody bindings, major scripting improvements(prep for full serialization), compiles on windows, a lot of editor improvements (can now display a lot more things from C# scripts).

+++ Crowny/Source/Crowny/ImGui/ImGuiVulkanLayer.cpp
@@ -36,7 +36,7 @@ namespace Crowny
         poolCreateInfo.sType = VK_STRUCTURE_TYPE_DESCRIPTOR_POOL_CREATE_INFO;
         poolCreateInfo.flags = VK_DESCRIPTOR_POOL_CREATE_FREE_DESCRIPTOR_SET_BIT;
         poolCreateInfo.maxSets = 1000;
-        poolCreateInfo.poolSizeCount = std::size(pool_sizes);
+        poolCreateInfo.poolSizeCount = (uint32_t)std::size(pool_sizes);
         poolCreateInfo.pPoolSizes = pool_sizes;
 
         VkResult result = vkCreateDescriptorPool(gVulkanRenderAPI().GetPresentDevice()->GetLogicalDevice(),
